add regex help for \* wildcards
