+++ app/src/state/actions/column-actions.ts
@@ -38,6 +38,8 @@ import {
 import { cloneDeep } from "lodash";
 import {
   DataMiningStatisticApproach,
+  EventSearchInfo,
+  GroupedEventSearchInfo,
   WindowStats,
   convertDataMiningPointDataTypeToDataMiningStatisticApproach
 } from "../../types";
@@ -46,7 +48,7 @@ import {
   computeWindowStatisticsForDataPoints,
   findRangeOfWindowStats
 } from "../../util/data-mining";
-import { yieldControl } from "../../util";
+import { getRegex, yieldControl } from "../../util";
 import { altUnitNamePrefix } from "../../util/constant";
 import { findSerialNum } from "../../util/util";
 
@@ -439,7 +441,6 @@ export const setColumnSelected = action((name: string) => {
 
 export const searchColumns = action(async (searchTerm: string, counter = { count: 0 }) => {
   setColumnSearchTerm(searchTerm);
-  await yieldControl(counter, 30);
   if (searchTerm === "") {
     state.settingsTabs.columnHashMap.forEach((columnInfo) => {
       setExpanded(false, columnInfo);
@@ -452,13 +453,15 @@ export const searchColumns = action(async (searchTerm: string, counter = { count
     return;
   }
   for (const columnInfo of state.settingsTabs.columnHashMap.values()) {
+    await yieldControl(counter, 30);
     setShow(false, columnInfo);
     setExpanded(false, columnInfo);
   }
 
-  const regExp = new RegExp(searchTerm);
+  const regExp = getRegex(searchTerm);
 
   for (const columnInfo of state.settingsTabs.columnHashMap.values()) {
+    await yieldControl(counter, 30);
     const match =
       columnInfo.name.toLowerCase().includes(searchTerm.toLowerCase()) ||
       regExp.test(columnInfo.name) ||
@@ -786,6 +789,122 @@ export const addAgeColumn = action((column: ColumnInfo) => {
   });
   column.children.splice(column.children.length, 0, ageColumn);
   state.settingsTabs.columnHashMap.set(ageColumnName, ageColumn);
+});
+export const makeColumnPath = action((name: string): string[] => {
+  const columnPath: string[] = [];
+  let column = state.settingsTabs.columnHashMap.get(name);
+  if (!column) {
+    return [];
+  }
+  while (column.name !== "Chart Root") {
+    columnPath.push(column.editName);
+    column = state.settingsTabs.columnHashMap.get(column.parent!);
+    if (!column) break;
+  }
+  return columnPath;
+});
+export const searchEvents = action(async (searchTerm: string, counter = { count: 0 }) => {
+  setEventSearchTerm(searchTerm);
+  let count = 0;
+  if (state.settingsTabs.eventSearchTerm === "") return { groupedEvents: [], count: 0 };
+
+  //key: column name/event name
+  //info: info found in subinfo array
+  const results = new Map<string, EventSearchInfo[]>();
+
+  for (const columnInfo of state.settingsTabs.columnHashMap.values()) {
+    await yieldControl(counter, 30);
+    if (columnInfo.name === "Chart Root") {
+      continue;
+    }
+    if (columnInfo.editName.toLowerCase().includes(state.settingsTabs.eventSearchTerm.toLowerCase())) {
+      //for column names
+      const id = columnInfo.editName + " - " + "Column";
+      if (!results.has(id)) {
+        results.set(id, []);
+      }
+      results.get(id)!.push({
+        id: count,
+        columnName: columnInfo.name,
+        columnPath: makeColumnPath(columnInfo.name),
+        unit: columnInfo.units
+      });
+      count++;
+    }
+    if (columnInfo.subInfo) {
+      //skip since subInfo is not associated with this but the app does it for map points
+      if (columnInfo.columnDisplayType === "MetaColumn") {
+        continue;
+      }
+      for (let i = 0; i < columnInfo.subInfo.length; i++) {
+        const subInfo = columnInfo.subInfo[i];
+        if ("label" in subInfo && subInfo.label) {
+          const regExp = getRegex(state.settingsTabs.eventSearchTerm);
+          if (regExp.test(subInfo.label)) {
+            const resultType = columnInfo.columnDisplayType === "Zone" ? "Block" : columnInfo.columnDisplayType;
+            const resInfo: EventSearchInfo = {
+              id: count,
+              columnName: columnInfo.name,
+              columnPath: makeColumnPath(columnInfo.name),
+              unit: columnInfo.units
+            };
+
+            //facies/chron label doesn't have subinfo because they are block type but its parent has facies/chron info, so access it through BlockSeriesMetaColumn
+            if (columnInfo.columnDisplayType === "BlockSeriesMetaColumn") {
+              if (state.settingsTabs.columnHashMap.get(columnInfo.name + " Facies Label")) {
+                resInfo.columnPath = makeColumnPath(columnInfo.name + " Facies Label");
+                resInfo.columnName = columnInfo.name + " Facies Label";
+              } else if (state.settingsTabs.columnHashMap.get(columnInfo.name + " Chron Label")) {
+                resInfo.columnPath = makeColumnPath(columnInfo.name + " Chron Label");
+                resInfo.columnName = columnInfo.name + " Chron Label";
+              } else {
+                console.error(
+                  "While searching, could not find Facies or Chron label for " +
+                    columnInfo.name +
+                    " but should have found it"
+                );
+                continue;
+              }
+            }
+            if ("age" in subInfo) {
+              //facies and chron label show up as block, so find ranges for them too
+              if (resultType === "Block" || columnInfo.columnDisplayType === "BlockSeriesMetaColumn") {
+                if (i > 0) {
+                  const nextBlock = columnInfo.subInfo[i - 1];
+                  if ("age" in nextBlock) resInfo.age = { topAge: nextBlock.age, baseAge: subInfo.age };
+                } else resInfo.age = { topAge: subInfo.age, baseAge: subInfo.age };
+              } else {
+                resInfo.age = { topAge: subInfo.age, baseAge: subInfo.age };
+              }
+            }
+            if ("subEventType" in subInfo) {
+              resInfo.qualifier = subInfo.subEventType;
+            }
+            if ("popup" in subInfo) {
+              resInfo.notes = subInfo.popup;
+            }
+            //same special case as above
+            const key =
+              resultType === "BlockSeriesMetaColumn"
+                ? subInfo.label + " - " + "Block"
+                : subInfo.label + " - " + resultType;
+            if (!results.has(key)) {
+              results.set(key, []);
+            }
+            const eventGroup = results.get(key)!;
+            eventGroup.push(resInfo);
+            count;
+          }
+        }
+      }
+    }
+  }
+
+  const groupedEvents: GroupedEventSearchInfo[] = [];
+  results.forEach((info: EventSearchInfo[], key: string) => {
+    groupedEvents.push({ key: key, info: [...info] });
+  });
+  return { groupedEvents, count };
 });
 
 export const changeAgeColumnJustification = action((column: ColumnInfo, newJustification: "left" | "right") => {
